requirements.txt
amqp==5.0.9
asgiref==3.4.1
billiard==3.6.4.0
celery==5.2.3
cffi==1.15.0
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cryptography==3.4.7
DateTime==4.3
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2
django-appconf==1.0.4
django-bootstrap-datepicker-plus==3.0.5
django-celery-beat==2.2.1
django-crispy-forms==1.12.0
django-cryptography==1.0
django-environ==0.4.5
django-on-heroku==1.1.2
django-timezone-field==4.2.3
et-xmlfile==1.1.0
gunicorn==20.1.0
honcho==1.1.0
kombu==5.2.3
openpyxl==3.0.7
packaging==21.3
Pillow==8.4.0
prompt-toolkit==3.0.24
psycopg2-binary==2.9.1
pycparser==2.20
pyparsing==3.0.7
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.3
redis==4.1.2
six==1.16.0
sqlparse==0.4.1
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
wrapt==1.13.3
XlsxWriter==1.4.4
zope.interface==5.4.0
